- More working on virtual memory
git-svn-id: svn://localhost/mosa@413 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/Kernel/Memory/IPhysicalPageManager.cs
@@ -33,7 +33,7 @@ namespace Mosa.Kernel.Memory
 		/// <summary>
 		/// Retrieves the size of a single memory page.
 		/// </summary>
-		ulong PageSize { get; }
+		uint PageSize { get; }
 
 		/// <summary>
 		/// Retrieves the amount of total physical memory pages available in the system.
